+++ CHANGELOG.md
@@ -10,6 +10,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 - #420: SSH tunneling implemented
 ### Changed
 - #1767: Turned about window into a simple popup form
+- #1766: Converted components check page into options page
 - #1460: Updated GeckoFX to v60
 ### Fixed
 - #1610: Menu bar changes to english when cancelling options form